f0 can be a double after function call

+++ tests/end_to_end/float-fn/irix-o2-out.c
@@ -1,4 +1,4 @@
 f32 test(void)
 {
-    return (f32) (f64) func_004000C4((f64) func_004000BC());
+    return (f32) func_004000C4((f64) func_004000BC());
 }
